+++ src/lib/py/subcommands.py
@@ -1,7 +1,11 @@
 import Convert2MzML
 import DownloadData
+import PhilosopherDatabase
+import RunFragPipe
 
 SUBCOMMANDS = [
     Convert2MzML.Convert2MzML,
     DownloadData.DownloadData,
+    PhilosopherDatabase.PhilosopherDatabase,
+    RunFragPipe.RunFragPipe,
 ]