+++ CHANGELOG.md
@@ -7,9 +7,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ---
 
+## 2026-01-07
+
+- Add script PhilosopherDatabase and RunFragPipe
+
 ## 2026-01-06
 
-- Add command DoenloadData
+- Add script DownloadData
 
 ## 2025-12-22